requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
python-multipart==0.0.9
jinja2==3.1.5

# VectorScan pipeline
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.0.0
pikepdf>=8.0.0
vtracer>=0.6.0
